Commit after updating the Second Image

+++ PiLOT_G2_IMG2/PiLOT_sb_MSS_CM3_app/Includes/RS485_Controller.h
@@ -36,6 +36,6 @@
 #define PAY_ID	0x07
 #define CLKS_PER_BIT    MSS_SYS_M3_CLK_FREQ / RS485_BAUD_RATE
 
-uint16_t init_RS485_Controller();
+uint8_t init_RS485_Controller();
 
 #endif /* INCLUDES_RS485_CONTROLLER_H_ */
